+++ packages/dataviews/src/hooks/index.ts
@@ -1,4 +1,4 @@
 /**
  * Internal dependencies
  */
-export { default as useIsFormValid } from './use-is-form-valid';
+export { default as useFormValidity } from './use-form-validity';
