+++ packages/dataviews/src/index.ts
@@ -2,6 +2,6 @@ export { default as DataViews } from './components/dataviews';
 export { default as DataViewsPicker } from './components/dataviews-picker';
 export { default as DataForm } from './components/dataform';
 export { default as filterSortAndPaginate } from './utils/filter-sort-and-paginate';
-export { useIsFormValid } from './hooks';
+export { useFormValidity } from './hooks';
 export { VIEW_LAYOUTS } from './dataviews-layouts';
 export type * from './types';
